+++ packages/core/src/renderer/components/+namespaces/namespace-storage.injectable.ts
@@ -3,7 +3,7 @@
  * Licensed under MIT License. See LICENSE in root directory for more information.
  */
 import { getInjectable } from "@ogre-tools/injectable";
-import clusterFrameContextForNamespacedResourcesInjectable from "../../cluster-frame-context/for-namespaced-resources.injectable";
+import allNamespacesInjectable from "../../cluster-frame-context/all-namespaces.injectable";
 import { isDefined } from "../../utils";
 import createStorageInjectable from "../../utils/create-storage/create-storage.injectable";
 
@@ -11,10 +11,10 @@ const selectedNamespaceStorageInjectable = getInjectable({
   id: "selected-namespace-storage",
   instantiate: (di) => {
     const createStorage = di.inject(createStorageInjectable);
-    const context = di.inject(clusterFrameContextForNamespacedResourcesInjectable);
-    const defaultSelectedNamespaces = context.allNamespaces.includes("default")
+    const allNamespaces = di.inject(allNamespacesInjectable);
+    const defaultSelectedNamespaces = allNamespaces.get().includes("default")
       ? ["default"]
-      : [context.allNamespaces[0]].filter(isDefined);
+      : [allNamespaces.get()[0]].filter(isDefined);
 
     return createStorage("selected_namespaces", defaultSelectedNamespaces);
   },
